+++ CHANGELOG.md
@@ -21,11 +21,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Checks to prevent players from being teleported in the ground
 - Checks on certain server commands to prevent players from being teleported into the ground 
 - Hunters now function as traders and sell weapons
+- `GradientBrick` fill allows structures to have a brick pattern while also having a gradient between two colors
 
 ### Changed
 
 - Can now swim through lillypads.
 - Music changes when entering a new site.
+- Flat bridge exteriors are gradients from surface block color to the construction stone color
 
 ### Removed